requirements.txt
pytest==8.0.0
selenium==4.17.2
requests==2.31.0
jsonschema==4.21.1
pytest-dotenv==0.5.2